@@ -23,11 +23,11 @@ const (
 	PaddingBlockSize       = 128 // RFC8467 recommendation
 )
 
-const kOptRrHeaderLen int = 1 + // DOMAIN NAME
-	2 + // TYPE
-	2 + // CLASS
-	4 + // TTL
-	2 // RDLEN
+// const kOptRrHeaderLen int = 1 + // DOMAIN NAME
+// 	2 + // TYPE
+// 	2 + // CLASS
+// 	4 + // TTL
+// 	2 // RDLEN
 
 const kOptPaddingHeaderLen int = 2 + // OPTION-CODE
 	2 // OPTION-LENGTH
@@ -68,7 +68,6 @@ func AddEdnsPadding(rawMsg []byte) ([]byte, error) {
 		switch body := additional.Body.(type) {
 		case *dnsmessage.OPTResource:
 			optRes = body
-			break
 		}
 	}
 	if optRes != nil {
