@@ -30,7 +30,7 @@ import (
 // separately closed.  (UNIX only.)
 func MakeTunFile(fd int) (*os.File, error) {
 	if fd < 0 {
-		return nil, errors.New("Must provide a valid TUN file descriptor")
+		return nil, errors.New("must provide a valid TUN file descriptor")
 	}
 	// Make a copy of `fd` so that os.File's finalizer doesn't close `fd`.
 	newfd, err := unix.Dup(fd)
@@ -39,7 +39,7 @@ func MakeTunFile(fd int) (*os.File, error) {
 	}
 	file := os.NewFile(uintptr(newfd), "")
 	if file == nil {
-		return nil, errors.New("Failed to open TUN file descriptor")
+		return nil, errors.New("failed to open TUN file descriptor")
 	}
 	return file, nil
 }
